Basic implementation of new API
Factor bucket methods into their own module and
update the tests to the new API.

There are 2 failing tests, one due to as CAS issue
and the other due to something about how connecting
to Couchbase changes the exception handling model
of the entire runtime.

Once these are resolved I think this library is
ready to share with the world.

Change-Id: I8e911a9d8023a979354102eaff37482b0fe521bf
Reviewed-on: http://review.couchbase.org/21456
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>