Fix the default configuration
People shouldn't be using Administrator for the instance, but
auth for the correct bucket. One may access the default bucket
without any credentials so we don't need a default user.
The port 9000 is only used from "cluster_run" that is used
internally from people developing on the couchbase server, and
should not be used by default by anyone else

Change-Id: I96168cfb851dff22abd379aeca8e21d2632911f3
Reviewed-on: http://review.couchbase.org/22798
Tested-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>