Add a couchnode benchmark sample
A work in progress for benchmarking the node client library. This commit
features measuring throughput and min, max, avg, 95 percentile response
times. The number of connections and requests are configurable through
command line as  well as on code(bench_config object). Additionally an
optional ramp up time is also configurable in the code. Currently
supports GET and SET ops.

Change-Id: Ic639b4673c59a7c85cdafe639bccd54fd1506aed
Reviewed-on: http://review.couchbase.org/26521
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>