JSCBC-49: Implemented overrides for cross-library compatibility.
Updated get to take a format option rather than simply raw to allow
the user to specify a specific format that they wish to retrieve the data
as.  Additionally, the set method has been modified to allow overriding
of the flags values instead of it being automatically determined by
the library.

Change-Id: Ia4882beee6bbec719f89089011b6596bcdc40ae5
Reviewed-on: http://review.couchbase.org/28849
Tested-by: Brett Lawson <brett19@gmail.com>
Reviewed-by: Mark Nunberg <mnunberg@haskalah.org>