Replace mentions of 'meta'
This nomenclature was sadly passed on from the previous versions of the
client. In the new API there is no "meta", but only a single result
object which contains both the value and any associated metadata.

Mentions of meta have been replaced with 'options' or 'result' where
applicable. Mentions of 'meta' are still within the views tests where
this 'meta' refers to an argument passed to the server-side view
function and is not related to the SDK terminology.

Change-Id: I7acc028d694ed4894a22235d9213742e995817dc
Reviewed-on: http://review.couchbase.org/29417
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Brett Lawson <brett19@gmail.com>