Do not JSON.parse id_range in MockBucket._execView
startkey_docid and endkey_docid are either undefined, or strings containing a doc id. As such, they shouldn't be parsed. This is already the behavior of the real implementation. The mock implementation is failing my tests, although it works with the real one.

The change for group_level is only a simplification in JS.