prebuild for linux and macos
For some reason, this project uses prebuilds but not prebuild-install
to download the builds from GitHub. This commit solves the issue
for *nix platforms (windows support to come at a later time).

Note that the repository in `package.json` has been changed to
my personal one, so that I may test building on my own repo.