JSCBC-377: Propagate secondary N1QL querying errors.
Previously, secondary errors (errors received after result rows) were
not propagated to the error parameter of the callback function.  This
meant you may get incomplete result sets with little indication of the
failure (as it would only show in the meta).  This change causes these
errors to show up as normal error objects on the callback.

Change-Id: I78fcc38068376846367ce1ec29d90eec8563366d
Reviewed-on: http://review.couchbase.org/83468
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>