JSCBC-392: Add unparseable error data to error message.
Motivation
----------
On some rare occasions, the answer from the cluster may not be fully parseable as JSON.  In these cases, completely ignoring the data causes the error to be extrodinarily difficult to debug.

Changes
-------
If we are unable to parse the response data as JSON, we append it to the error message as a string.

Change-Id: Ia0bd94fd6aa8171b0542194ac415f7dc5ba17910
Reviewed-on: http://review.couchbase.org/80460
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>