JSCBC-396: Fix getIndexes to return only indexes of the target bucket.
Motivation
----------
A user identified that the `getIndexes` method of the bucket manager was
returning all indexes for the cluster, rather than only the indexes for
the target bucket.  It also failed to correctly order the returned indexes
in a manner that would allow them to be built in-order (N1QL requires the
primary index to be build first).

Changes
-------
Added additional predicates to the N1QL query used to fetch only the
indexes for the target bucket. It additionally now includes ordering
parameters to ensure that primary indexes are returned first.

Change-Id: Ifd3e643f0758a1739f443158b7d673b8a2b030d3
Reviewed-on: http://review.couchbase.org/80537
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>