JSCBC-400: Don't include empty data in management errors.
Motivation
----------
Due to a recent change to the error message formatting for
management errors, the error message included an additional
comma at the end of the error when the response body was
empty.  This looked strange and caused one of the unit
tests to fail.

Changes
-------
We now validate that the response body was not blank before
we attempt to include it in the error message.

Change-Id: Ib9a2da8fdb9d54e67f951937d57996de63b22174
Reviewed-on: http://review.couchbase.org/81768
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>