JSCBC-403: GetReplica test should not depend on node count.
Motivation
----------
The GetReplica test previously would work or not work depending
on the size of the cluster we were testing on.  In the future
we should perform checks for this, but at the moment we will
simply make it ignore node counts.

Changes
-------
Pass an impossible node index to the GetReplica function in the
GetReplica test.

Change-Id: I4cb0507ff12849a5972434b12103d742fa600e10
Reviewed-on: http://review.couchbase.org/82323
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>