JSCBC-424: Correctly check there is meta-data before deleting errors.
Motivation
----------
A user has discovered an edge-case that causes an exception to be
thrown within the Node.js SDK.  This case occurs when the N1QL service
responds with invalid JSON.  The SDK attempts to remove the errors
property from an invalid object, throwing an exception.

Changes
-------
Moved the data to its own variable for clarity on the purpose, as well
as check that there is actually data to delete prior to removing it.

Change-Id: I1e25f2e02e5b9ccd3c363fdeac4e0da885672a5a
Reviewed-on: http://review.couchbase.org/84274
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>