JSCBC-428: Add handler for HTTP request errors.
Motivation
----------
Previously, the SDK would fail to register for the error
event from the HTTP requests that it prepares.  This would
cause the error to be dispatched via the uncaughtException
handler in Node.js which is incorrect behaviour.

Changes
-------
Catch the errors for each HTTP request and forward the
errors through the normal callback path.

Change-Id: I974413800d06f829e06efca2a52232cc370a6625
Reviewed-on: http://review.couchbase.org/85029
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>