JSCBC-439: Added method to fetch top-level couchbase library from bucket.
Motivation
----------
Due to the fact that numerous object types that need to be passed into
the bucket methods hang off the couchbase library itself, it is
impossible to create the correct object types without a reference
to an identical couchbase library that created the bucket.

Changes
-------
Expose a new cbLib property on the bucket that allows you to create
query objects.

Change-Id: If2ab601cbbae87712fe193859e5cf846664566cc
Reviewed-on: http://review.couchbase.org/86395
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>