JSCBC-452: Don't immediately close connecting buckets.
Motivation
----------
Attempting to disconnect a bucket before the connection phase
completed would cause the SDK to segfault, as it would attempt
to begin asynchronously cleaning up resources that were still
in the process of being constructed.

Changes
-------
Modified the disconnect method to run through the typical
invocation wait queue as normal operations to ensure that the
connection has been established before we try to disconnect it.
(and to ensure that disconnect calls do eventually get handled).

Change-Id: I5a8058e0dccb9c77b2c5183601319726988b2f6f
Reviewed-on: http://review.couchbase.org/90090
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>