JSCBC-471: Fixed issue where instances were double-destructed.
Motivation
----------
Under some circumstances (which are readily apparent in tests),
the SDK will allow certain calls to be performed on destroyed
Couchbase instances.  This also caused a double-destruction of
the instance when the Node.js object was latter cleaned up.

Changes
-------
Removed lcb_destroy_async from the onConnect callback.  This
leaves the instance disconnected but intact, allowing
operations to fail as expected as opposed to triggering
a segfault or other memory access error.

Change-Id: I4a90eb69ca372d5c4f3facc571f49b6f014ae8a0
Reviewed-on: http://review.couchbase.org/93256
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>