JSCBC-472: Updated libcouchbase version restriction.
Motivation
----------
In order to make it more obvious why compilation may be
failing, there are checks in place to ensure that an
appropriate libcouchbase version is used with the SDK.
This check has not been updated recently and should be.

Changes
-------
Updated libcouchbase version compatibility check to
match with the built-in library version.

Change-Id: I6e7e765c91f2982a825b9b26f64d0dc2f188704c
Reviewed-on: http://review.couchbase.org/93260
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Brett Lawson <brett19@gmail.com>