JSCBC-476: Error when trying to mix authentication types.
Motivation
----------
The use of a cluster-level authenticator alongside a bucket
password does not make a lot of sense.  We should be throwing
an error in this case to force the user to clarify which
is intended to be used.

Changes
-------
We now check if both an authenticator and bucket password
were specified when opening a bucket and throw an error if
this is the case.

Change-Id: I53bf337dcd2b1ecd1336daf574d762098c06d5be
Reviewed-on: http://review.couchbase.org/94543
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>