JSCBC-492: Fixed unnecessary warnings when no loggers are active.
Motivation
----------
When all logging is explicitly disabled (for instance, when
tests are being performed).  The internal winston instance
will log warnings to the console due to having no loggers to
send messages to.

Changes
-------
Added a special case where, when no loggers are enabled for
winston, we attach a special logger which consumes all log
messages without emitting anything.

Change-Id: I550cb278c7dc30d8eb01a310efc5be95f1063ed2
Reviewed-on: http://review.couchbase.org/97419
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>