JSCBC-497: Added integrated support for Couchbase Analytics.
Motivation
----------
The previous implementation of analytics in this SDK relied on the
user providing the hostnames for a Couchbase Analytics service and
then generating the requests manually on our side.  Now that
analytics is a first class service of the Couchbase platform, this
should be modified to use the information available in our configs.

Changes
-------
Refactored our analytics querying code to take advantage of the
built-in support in libcouchbase.

Change-Id: I4471e364a5e55bb4c18aa57c22211f3698e9ada0
Reviewed-on: http://review.couchbase.org/100145
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>