JSCBC-499: Removed unused options docs from lookupIn.
Motivation
----------
All methods in the SDK have the plumbing to support an options
object, but in the case of lookupIn, there were no publicly
exposed options available.  This occasionally causes confusion
with users expecting that the object has properties.

Changes
-------
Removed the documentation item for the options object on the
lookupIn method of the Bucket object.

Change-Id: I0c9e88c9559ddec7b3fb62ea6bfafbd30563f1c0
Reviewed-on: http://review.couchbase.org/100206
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>