JSCBC-515: Fix special characters in versions corrupting requests.
Motivation
----------
The most recent build of Node.js has included special characters
in the version numbers that it exposes to applications.  This data
is forwarded to libcouchbase which then fails to correctly generate
a user-agent string header.

Changes
-------
Added trimming to all version numbers used when generating a
user-agent string for libcouchbase.

Change-Id: I6fc06955e64b4e010dface6c0b2ddd55b06fda04
Reviewed-on: http://review.couchbase.org/101358
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>