Replace `v8::Handle` with `v8::Local`
Back in 2015, in the v8/v8 repository, commit `202a97c` made `Handle`
a synonym of `Local` and then commit `f022e53` started the formal
deprecation process of `v8::Handle`. The synonym change first landed
in V8 v4.4.27 and the `V8_DEPRECATE_SOON` change was released
as V8 v4.5.98.

In accordance to the version table of Node.js, this means
the deprecation was released with Node.js v4.0.0 (being released
with V8 v4.5.103.30). Considering Node.js v4 became End-of-life
on 2018-04-30, it is safe to carry out the change for all supported
Node.js versions.

Ref: https://chromium.googlesource.com/v8/v8.git/+/f022e53b9579457d53ee59b46ce6a959d27669a4
Ref: https://chromium.googlesource.com/v8/v8.git/+/202a97c88eb6807915135a1e6be9aa921cd8efac
Change-Id: Iddf8f5f6ff41dbb8cbdd0613ad336cb0dc1583ba
Reviewed-on: http://review.couchbase.org/110068
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>