Pass a context for V8 transformations
Back in 2015, v8/v8 introduced `5234d99` which added a maybe version
of `FunctionTemplate::GetFunction` along with a `V8_DEPRECATE_SOON`
on the pre-existing `GetFunction` method. In 2019 commit `6f5e805`
dropped that method entirely. As the initial 'deprecate soon' landed
with V8 v4.3.31 (that is io.js v3.0.0 prior to Node.js v4!) it is
fairly safe to switch to the maybe version. In addition to that,
the full removal landed with V8 v7.4.66, which is being shipped
with Node.js v12 now, breaking all builds for the upcoming LTS.

Similarly, `Value::ToString` received a maybe version back in
2015 with commit `2fc51d9` followed by a `V8_DEPRECATE_SOON`
with `47154aa` which released with V8 v4.3.27 (io.js v3.0.0).
Honestly not sure what commit dropped the method, but it's not
there anymore either for Node.js v12.

Change-Id: Iac4012b31f4f5b747bccdc3ec079b9d615479951
Reviewed-on: http://review.couchbase.org/110167
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>