Rework the `CONTRIBUTING` documentation
The document there is ages old, referencing a `manifest.xml` that has
ben unmaintained for a while already. While it can be a solid entry
barrier for contributors, as in order to filter out the ones who are
'dedicated enough' to figuring it out, it does feel like it could be
confusing to anyone new to the project who is not sure where to start.

Change-Id: Ie872a1afc51e35319abe4d53a86460f32eb8b940
Reviewed-on: http://review.couchbase.org/110494
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>