JSCBC-581: Corrected incorrect documentation for FTS spatial query.
Motivation
----------
The documentation originally indicated that the distance should
be specified as a numeric value, when it should instead indicate
that the value is specified as a string which includes the units.

Changes
-------
Modified documentation to indicate a string rather than number
should be passed for the distance on GeoDistanceQuery.

Change-Id: Ie232f1bfbd9063a0addf6dd3af0037e5084bbc01
Reviewed-on: http://review.couchbase.org/113836
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Brett Lawson <brett19@gmail.com>