JSCBC-589: Correct geospatial lon/lat ordering.
Motivation
----------
Due to an oversight during the implementation of FTS
geo-spatial querying in the Node.js SDK, the parameter
order was defined as lat,lon and sent as lat,lon when
the server expected the point as lon,lat.

Changes
-------
As this error occurred many years ago, and in the
interest of maintaining compatibility across the
existing customer base, we have appropriately renamed
and re-documented the parameters to reflect the true
ordering as implemented rather than as defined.

Change-Id: I612c58554ed8fda1710cc67d7bbb461fd7f41f9e
Reviewed-on: http://review.couchbase.org/114652
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>