JSCBC-613: Added support for full-document fallback on projections.
Motivation
----------
In order to provide a more sensible user experience, the SDK is
intended to automatically fall back to fetching the full document
if there are too many fields being projected to use subdocument.

Changes
-------
Added handling for fetching the full document when the number of
fetch fields exceeds the server limit.  Additionally added a system
to reproject the document such that only fields requested are
actually returned.

Change-Id: I72fd41902834ca256ce63d4e6c32a4f33d3b762c
Reviewed-on: http://review.couchbase.org/119616
Reviewed-by: Matt Ingenthron <ingenthr@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>