JSCBC-630: Corrected issue with decoding invalid JSON.
Motivation
----------
The SDK is intended to fall back to returning a Buffer object
whenever JSON data is not parseable.  Currently this actually
causes the SDK to fail to throw a semi-fatal exception.

Changes
-------
Improved the error-handling code now that we utilize the JSON
parsing facilities of Nan as opposed to using v8 itself.

Change-Id: I83d0587199382411a47854f9ca84beb8bdf63e55
Reviewed-on: http://review.couchbase.org/119509
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>