JSCBC-668: Fixed promise reject error when using callbacks.
Motivation
----------
The SDK is intended to support both callback-oriented programming
as well as async/await.  Currently the use of callbacks can trigger
unhandled promise rejections due to the inclusion of promises even
when callbacks are used.

Changes
-------
Updated the promise wrapping code to generate either promises or
callback handlers, but not both.

Change-Id: I969f04eccf9a9b09e6eb3e4c8aaace1186df940c
Reviewed-on: http://review.couchbase.org/123225
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Tested-by: Brett Lawson <brett19@gmail.com>