JSCBC-68: Fixed metrics microsecond parsing.
Motivation
----------
Times in metrics that are in microseconds are converted as
though they are in milliseconds.

Changes
-------
The issue is that if there is not a conversion factor for
parse[ units ], a conversion factor of 1 is used.
The 'μs' built-in to parse-duration is not the same 'μs'
used by couchbase.
When utils.js uses parseDuration, add the 'μs' that
couchbase uses to the parsing data structure

Change-Id: Ic80318f2c2dca11f4c05495d12c31c20727d5cf2
Reviewed-on: http://review.couchbase.org/122626
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Tested-by: Brett Lawson <brett19@gmail.com>