JSCBC-711: Fixed issue with early shutdown of flush watcher.
Motivation
----------
Due to the placement of our setup code, it was possible for
the SDK to attempt to clean up our flush watcher before it
was set up, causing a segfault.

Changes
-------
Moved the flush watcher initialization to the constructor
of the Connection rather than being in the connect callback.

Change-Id: I840d70a8fea6d5029233c1f028b736355e8f542d
Reviewed-on: http://review.couchbase.org/124052
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Tested-by: Brett Lawson <brett19@gmail.com>