JSCBC-676: Fixed view query to return proper row object.
Motivation
----------
The view query API is intended to return a row object which
contains the row value, document id and the index key.  Due
to a missunderstanding of the libcouchbase API however, the
row emitter was only emitting the row value directly.

Changes
-------
Modified the row emitter to emit the row object rather than
the row value directly.

Change-Id: I69b5edb3f1a3d76c22fb8a9bb18187f0262d092f
Reviewed-on: http://review.couchbase.org/c/couchnode/+/127258
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>