JSCBC-729: Added roles serialization to users.
Motivation
----------
An oversight in the implementation of the user management
API caused User objects not to serialize their roles when
upserting to the user.

Changes
-------
Added roles to the fields which are serialized from a User
object before being sent to the server.  Enabled testing of
user management on 6.0.0 clusters.

Change-Id: I86dedeeebaba120694f14d6cbbc21b79ed0e84aa
Reviewed-on: http://review.couchbase.org/c/couchnode/+/127532
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>