JSCBC-730: Adjusted LookupIn to return content not results.
Motivation
----------
The RFC defining SDK behaviour for lookupIn operations indicates
that the proper name for the results of the lookupIn operation
should be called 'content' rather than 'results'.

Changes
-------
Adjusted the return object to containing a content property rather
than results property.  Added a backwards-compatibility definition
which enables the use of results still.

Change-Id: Ib6578e2553b62b342ed54a10e50db6b39cff39bb
Reviewed-on: http://review.couchbase.org/c/couchnode/+/127533
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>