JSCBC-754: Don't swap in a bucket name when none was used.
Motivation
----------
When internally establishing cluster-level connections, the
connection string we use was being passed to the connection
string parsing code to handle options insertion.  This was
causing a bucket name to be inserted when it should not have.

Changes
-------
Modified connection string management to avoid inserting a
bucket name if there was not one parsed.

Change-Id: Ic7e0d25ab6774eacbb407c36c0c60e415f82c0d5
Reviewed-on: http://review.couchbase.org/c/couchnode/+/130441
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>