JSCBC-758: Updated view scan consistency handling.
Motivation
----------
The scan consistency implementation in the SDK incorrectly
calls the field 'stale' as opposed to scan consistency.

Changes
-------
Updated the field name to scanConsistency rather than stale,
added an enumeration to cover the new field and implemented
a backwards-compatible handler for converting the stale
option to a ScanConsistency option.

Change-Id: I854f04637825d398a6cf0e0db6ececb1c97a506e
Reviewed-on: http://review.couchbase.org/c/couchnode/+/130583
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>