JSCBC-759: Fixed binary data being interpreted as UTF-8.
Motivation
----------
In some cases, the data being passed to the SDK comes in the
form of a Buffer object.  This was being parsed as a UTF-8
string, causing some types of buffers to be mangled when
inserted.

Changes
-------
Added explicit handling for Buffer objects and included an
additional test to cover binary data being used.

Change-Id: Ib0d06c2f42bd2dd1a9bbb07e396c71393ce8b1a6
Reviewed-on: http://review.couchbase.org/c/couchnode/+/130815
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>