JSCBC-788: Added high-level options for specifying certificates.
Motivation
----------
The SDK supports the use of TLS Certificate Authentication, but this
is only available through the use of connection string options directly.
Rather than requiring the use of the connection string, we should expose
high level options for this.

Changes
-------
Added the appropriate authenticator implementations as well as handling
for certificate authentication based authentictors.

Change-Id: I4475b9a6a78fbc4699e8a715317ce86d56ec3627
Reviewed-on: http://review.couchbase.org/c/couchnode/+/135388
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>