JSCBC-786: Added uncommitted collections support for user management.
Motivation
----------
Couchbase Server 7.0 will add full support for collections.  As part of
this, RBAC has been extended to allow access control on a per-scope or
per-collection basis and this should be accessible from the SDKs user
management facilities.

Changes
-------
Added support for scoping roles to a scope or scope and collection.

Change-Id: I13d196e4be627e495eca4b5f15cf974472f1e1e0
Reviewed-on: http://review.couchbase.org/c/couchnode/+/135391
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>