JSCBC-797: Resolved a number of typescript typings errors.
Motivation
----------
The shipped typings file contains errors making it unusable when
in some IDE's.

Changes
-------
Cleaned up a number of places that the documentation was incorrect
which lead to incorrect typings generation.  Additionally fixed an
issue where `make types` didn't put the file in the correct place,
and added a check to ensure the typings file will compile.

Change-Id: I9e17b8ebb028e8c0105eb88b2c2ea8bb4a8184c1
Reviewed-on: http://review.couchbase.org/c/couchnode/+/138776
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>