JSCBC-755: Added support for multi-value sub-document array ops.
Motivation
----------
According to the specification, arrayInsert/arrayAppend/arrayPrepend
are intended to take an array of values which should be inserted.
An oversight caused it to only accept a single value.

Changes
-------
Added an option called multi which enables the expected behaviour
where multiple values can be provided for the mentioned operations.

Change-Id: I4a88238b8a83a0598aa8c8cc5ec16d984c74fac0
Reviewed-on: http://review.couchbase.org/c/couchnode/+/138791
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>