JSCBC-803: Added support for pinging the cluster.
Motivation
----------
The 3.0 SDKs include the functionality to ping the cluster at both
the cluster and bucket level to validate that everything is
correctly connected and working.

Changes
-------
Fixed up the low level interface in the SDK and added the appropriate
high level methods which allow access to libcouchbase's ping method.

Change-Id: I0bc56280650112ff87e1c2dbc503bb0117a34ad0
Reviewed-on: http://review.couchbase.org/c/couchnode/+/139176
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>