JSCBC-761: Added support for specifying minimum bucket durability.
Motivation
----------
As of Couchbase Server 7.0, we now support specifying a minimum durability
level which is required to be satisfied for all mutation operations against
a bucket.  This should be exposed through the management API.

Changes
-------
Added ability to specify minimum bucket durability for the creation and
updating of a bucket in the management APIs.

Change-Id: I7b6a592d93d9bd69745e1cde1777a7bbe652e8ec
Reviewed-on: http://review.couchbase.org/c/couchnode/+/140819
Reviewed-by: Matt Ingenthron <ingenthr@gmail.com>
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>