JSCBC-815: Avoid re-using consumed va_list.
Motivation
----------
Due to the fact that using a va_list fundamentally alters it and
makes it no longer valid, we need to make a copy of the list in
order to enable us to call sprintf twice in the case that our
buffer needs to be resized.

Changes
-------
Modified logger to copy the va_list we receive before we invoke
the first sprintf call such that we still have a valid copy of
the va_list in case we need to invoke sprintf again.

Change-Id: I9ec650c934f3db49445d6d2d09facc0140bf2a9b
Reviewed-on: http://review.couchbase.org/c/couchnode/+/141256
Reviewed-by: Matt Ingenthron <ingenthr@gmail.com>
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>