JSCBC-816: Fix cluster errors not propagating for http methods.
Motivation
----------
When performing methods based on HTTP, there is a FINAL flag set when
its the final callback we expect.  When the SDK throws an error, it
calls the method with no flags at all.  This conflict causes the SDK
to throw a 'unexpected non-final callback' when the SDK itself throws
the error, such as when cluster connect has failed.

Changes
-------
Inverted the FINAL logic to instead use a NONFINAL flag.

Change-Id: Ic17458cd335cdc5c55a1e7f9ce633b3897dbe530
Reviewed-on: http://review.couchbase.org/c/couchnode/+/141264
Reviewed-by: Matt Ingenthron <ingenthr@gmail.com>
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>