JSCBC-856: Fixed memory leak with trace span management.
Motivation
----------
A memory leak was detected in the SDK by a user, after an
investigation into the source of the leak it was determined
that a copy-paste error lead to a second trace span being
created for each operation.  In addition, these trace spans
were not being cleaned up due to not being captured by our
scoped trace span wrapper.

Changes
-------
Fixed trace span creation to not be duplicated by an
incorrectly implemented call to the tracer.

Change-Id: I955530f675fbf53521b4f88483d285d5f658266f
Reviewed-on: http://review.couchbase.org/c/couchnode/+/148334
Tested-by: Brett Lawson <brett19@gmail.com>
Reviewed-by: Charles Dixon <chvckd@gmail.com>