JSCBC-858: Fixed issue where FeatureNotAvailable was thrown in error.
Motivation
----------
As part of the error handling for collection management we need to
check for 404 errors, but these can happen both when an entity does
not exist as well as when the feature isn't supported.  We should
check for the former before the latter.

Changes
-------
Moved the feature not available checking to the end of the error
handling.

Change-Id: I40559fa09befbf6ea7ac3e3fc21c7ecb9337662e
Reviewed-on: http://review.couchbase.org/c/couchnode/+/157893
Reviewed-by: Sergey Avseyev
Tested-by: Brett Lawson <brett19@gmail.com>