JSCBC-967: Use correct arguments indices for `options`
The `options` argument for many functions is optional.
It can be skipped (left `undefined`) or a faley value
such as `null` can be provided. These will all default
to an empty object with an `if (!options)` check.
The user of the SDK may also choose to provide a `callback`
function as an argument to these functions. As `options` is
optional, skipping it entirely is also allowed.

Simply put, providing `options = undefined` such as:
    collection.touch('mykey', 60, undefined, () => {})
should behave the same as skipping that argument:
    collection.touch('mykey', 60, () => {})

This works via a check, that if `options` is an instance of a Function,
then `callback` is set that value instead, and `options` gets cleared.
Due to historical reasons, `arguments` has been preferred.

When changing arguments lists for functions, one may simply forget
to tweak the arguments index which points at `options`. If an argument
is being removed and the index remains too high, then the `callback`
value will result in `undefined` and the callback will not be called.
If an argument is being added and the index remains too low, then the
`callback` will be set to an unexpected value. As in the case of
previously mentioned `collection.touch`, the value of `callback` would
be set to `60` (number) and later attempt to invoke `callback` fails
with a runtime error of 'callback is not a function'.

A change to allow `options` as both `Options` and `Callback` may
be beneficial, allowing `callback = options` to be used. However
due to my limited experience with Typescript, I am hesitant to propose
such a large scale rewrite. This may be favorable for maintainability,
as to not burden future copy-paste actions or argument list changes
with these index updates. As a result, this may be discussed, and
picked up by anyone comfortable with overhauling the function
signatures. This change instead tries to resolve the current behavior
bug while maintaining pre-existing structures.

Change-Id: I31cd32cfd2f751a4117e0e2ec67dfde9f3887df0
Signed-off-by: Valters Jansons <valter.jansons@gmail.com>
Reviewed-on: https://review.couchbase.org/c/couchnode/+/165690
Tested-by: Brett Lawson <brett19@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>