JSCBC-1094: Add missing transaction exception type handling.
Motivation
----------
At some point, the binding layer for transaction exceptions handling
was missed in a merge and is missing from the current binding.

Changes
-------
Added the missing binding code for the transaction exception types.
Additionally updated the list of enum values as new values have been
added since it was originally implemented.

Change-Id: I9e3138fa8ea3493bbbad5b0b5cfbbf77ba3ab95d
Reviewed-on: https://review.couchbase.org/c/couchnode/+/178611
Reviewed-by: Jared Casey <jared.casey@couchbase.com>
Tested-by: Brett Lawson <brett19@gmail.com>