JSCBC-1106:  Update SearchRow object's fields and explanation types
Motivation
----------
As much as possible we want to keep 3.x and 4.x response objects the same
to help ease migration efforts.  For search query responses, the SearchRow should have
the same object shape in both 3.x and 4.x versions of the SDK.

Changes
-------
Make sure a search query provides SearchRow objects that represent
fields and explanation as objects (if they exist in the response). Update tests to
confirm functionality.

Change-Id: Ifcea54e15aff66e7f8a2f5eb9b599ef0e4ef3cb7
Reviewed-on: https://review.couchbase.org/c/couchnode/+/185476
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Jared Casey <jared.casey@couchbase.com>