JSCBC-1160: WAN config profile should use system defaults for DNS nameserver and port
Motivation
----------
When using the WAN config profile we want to set the dns_srv_timeout and
leave the DNS nameserver and port at defaults.  However, the defaults we
want are the system config defaults and not the "fallback" defaults the
C++ client provides.

Changes
-------
Update C++ core to pull in DNS config changes and update how the WAN
config profile sets the default DNS config nameserver and port values.

Change-Id: I3e048cfd146d08311440a35ad8c42521213b565c
Reviewed-on: https://review.couchbase.org/c/couchnode/+/192394
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Jared Casey <jared.casey@couchbase.com>