JSCBC-1178: Provide access to protocol logger
Motivation
----------
The C++ core 1.0.0-dp.7 added support for a protocol logger (enables
network traffic to be saved to a file). Exposing the protocol logger
would be useful for deubgging in certain scenarios.  NOTE: this is a
volatile feature that is subject to future change/improvements.

Changes
-------
Expose C++ core protocol logger.

Change-Id: I04533e0e404316ab469eb2841e800f4ca60211a5
Reviewed-on: https://review.couchbase.org/c/couchnode/+/194753
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Jared Casey <jared.casey@couchbase.com>