JSCBC-1187: Fix connstr automatic override of trust_store_path
Motivation
----------
The SDK is automatically overriding the trust_certificate path with the
trustStorePath in the SecurityConfig even if the trustStorePath is
undefined.  This means users cannot pass in the trust_certificate via
the connection string. Also, the SDK does not translate the
trust_store_path query string param to trust_certificate (what the C++
client expects).

Changes
-------
Pick up trust_certificate from connection string and only override from
the SecurityConfig if the trustStorePath exists.  Also, allow use of the
trust_store_path in the connection string and translate to
trust_certificate.

Change-Id: I6b1faf495654f18b6036b06ef73ead11048b2d78
Reviewed-on: https://review.couchbase.org/c/couchnode/+/196568
Tested-by: Jared Casey <jared.casey@couchbase.com>
Reviewed-by: Mateusz <matt.wozakowski@couchbase.com>