JSCBC-1189: Fix crud tests that have callback
Motivation
----------
Tests with options and callback are failing on CI because their set timeout is too short, which isn't what is being tested. Additionally, the callback tests are not structured correctly causing the tests to look like they are failing with a mocha timeout.

Changes
-------
Increase timeout in options/callback tests. Restructure crud tests with callback

Change-Id: I51f958600adb9b7a68cfb72c29284e563870bca9
Reviewed-on: https://review.couchbase.org/c/couchnode/+/197748
Reviewed-by: Jared Casey <jared.casey@couchbase.com>
Tested-by: Mateusz <matt.wozakowski@couchbase.com>