JSCBC-1245: Refactor Document Expiry Duration
Motivation
----------
The changes for JSCBC-776 to refactor the document expiry did not carry
over to the 4.x series of the SDK.  We should update the SDK to include
the logic to refactor the document expiry duration.  The changes will
help get Node.js FIT to green.

Changes
-------
Add logic for handling document expiry.  Added tests to confirm changes.
Verify related FIT tests pass.

Change-Id: Id2774057772d5a4d0a01f95dab523b01f9593706
Reviewed-on: https://review.couchbase.org/c/couchnode/+/207550
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Jared Casey <jared.casey@couchbase.com>