JSCBC-1243: Unable to retrieve binary documents w/in transaction lambda
Motivation
----------
Even though the current transaction implementation of the SDK does not
support binary documents, the SDK should not crash in this scenario. How
the SDK translates the `TransactionGetResult` should look closer to how we
handle decoding JSON in the `DefaultTranscoder`.

Changes
-------
When translating the `TransactionGetResult` from C++, attempt to decode
the JSON and if that fails return the raw content.

Change-Id: I5bea63df8bee0745826181215f59be6292be3f54
Reviewed-on: https://review.couchbase.org/c/couchnode/+/208347
Reviewed-by: Mateusz <matt.wozakowski@couchbase.com>
Tested-by: Jared Casey <jared.casey@couchbase.com>